+++ CHANGES.md
@@ -2,6 +2,20 @@ CHANGES
 ========
 
 
+## 1.1.0
+
+* Feature: Use redis for session store!
+* Feature: Mail setting and added send mail module.
+* Feature: User invitation.
+* Feature: Activate invited user self (admin).
+* Feature: Activate registered user (admin / using RESTRICTED mode).
+* Feature: User suspend (admin).
+* Feature: User delete (admin).
+* Improve: Wiki style improved.
+* Improve: Update favicon (high resolution).
+* Fix: Affix header handling.
+* Library Update: Bootstrap 3.3.1, Fontawesome 4.2.0, async 0.9.0,
+
 ## 1.0.4
 
 * Feature: Basic auth restriction whole pages access.